redis: helper script in Python.
Wrote a small script to fetch an existing list from a Redis server,
add new quotes to it, shuffle the result and push it back to a new key,
with the option to rename the result to the original list,
thus replacing it.

This script could be rewritten in Rust but it is not particularly high
priority. I simply need it on occasion to "integrate" new quotes partway
through a posting cycle.